Fix possible invalid parsing config file
The "key_name" entry in config file is put in an array
separated by commas.
When there are some blanks around the comma, they are
not removed and space is included in array elements.
ex. "key_name aaa, bbb, ccc"=>["aaa", " bbb", " ccc"]
A correct array is: ["aaa", "bbb", "ccc"]
So fix it.